Added initial POC of tmdb download

+++ src/main/kotlin/com/github/couchtracker/server/common/model/Rating.kt
@@ -6,7 +6,7 @@ import kotlinx.serialization.Serializable
 
 sealed class Rating {
     @Serializable
-    data class Tmdb(val value: Float, val count: Long) : Rating()
+    data class Tmdb(val value: Double, val count: Long) : Rating()
 
     @Serializable
     data class Tvdb(val value: Float, val count: Long) : Rating()
